requirements.txt
pygame
neat-python
bresenham
